First cut of the OT-based `LocalFile` implementation.
As of this writing, this is super-duper inefficient, for at least a couple
reasons:

* `FrozenBuffer`s are encoded as base-64 strings.
* Change history is never culled, which means that nominally ephemeral changes
  never actually get removed from a file's history.
